requirements.txt
django
djangorestframework
rest_object=2.2.2
coveralls
djangorestframework-jwt
djoser
django-cors-headers